third_party/WebKit/LayoutTests/fast/dom/Selection/missing-arguments-expected.txt

PASS selection.addRange() threw exception TypeError: Failed to execute 'addRange' on 'Selection': 1 argument required, but only 0 present..
PASS selection.getRangeAt() threw exception TypeError: Failed to execute 'getRangeAt' on 'Selection': 1 argument required, but only 0 present..
PASS selection.selectAllChildren() threw exception TypeError: Failed to execute 'selectAllChildren' on 'Selection': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
